backend/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
flask-bcrypt>=1.0.1
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
